Refactor Windows App backend to have a common base for other formats.
